kuu nime võrdlus ja vigade parandus

+++ src/print_helper.h
@@ -2,6 +2,6 @@
 #define _PRINT_HELPER_H_
 
 int print_ascii_tbl (FILE *stream);
-int print_for_human (FILE *stream, const unsigned char *array, const size_t len);
+int print_for_human (FILE *stream, const unsigned char *array, const int len);
 
 #endif /* _PRINT_HELPER_H_ */
